+++ ddb/ddbpath/parse.go
